requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
bcrypt
psycopg2-binary
pydantic
